+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "nzcsa-frontend",
-  "version": "0.1.0",
+  "version": "3.1.0",
   "private": true,
   "dependencies": {
     "@3m1/service-worker-updater": "^2.0.3",
@@ -19,6 +19,7 @@
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
     "date-fns": "^2.0.0-beta.5",
+    "dotenv": "^16.0.0",
     "i": "^0.3.6",
     "jwt-decode": "^3.1.2",
     "material-ui-image": "^3.3.2",
