+++ src/pages/SignInSide.js
@@ -8,7 +8,6 @@ import Checkbox from '@material-ui/core/Checkbox';
 import Link from '@material-ui/core/Link';
 import Paper from '@material-ui/core/Paper';
 import Box from '@material-ui/core/Box';
-import GoogleIcon from '@mui/icons-material/Google';
 import Grid from '@material-ui/core/Grid';
 import LockOutlinedIcon from '@material-ui/icons/LockOutlined';
 import CircularProgress from '@material-ui/core/CircularProgress';
@@ -17,7 +16,7 @@ import { makeStyles } from '@material-ui/core/styles';
 import { Container } from '@material-ui/core';
 
 import { useState, useEffect } from 'react';
-import { login, signUp } from '../api/connectBackend';
+import { login } from '../api/connectBackend';
 
 import { red } from '@material-ui/core/colors';
 import RandomImagePicker from '../components/RandomImagePicker';
@@ -26,8 +25,7 @@ import { isMobile } from "react-device-detect";
 import Alert from "@material-ui/lab/Alert";
 import { isIos, isInStandaloneMode } from '../utils/pwaUtils';
 
-import Divider from '@mui/material/Divider';
-import { GoogleLogin } from 'react-google-login';
+import GoogleLoginButton from '../components/Auth/GoogleLoginButton';
 
 
 const useStyles = makeStyles((theme) => ({
@@ -155,53 +153,6 @@ export default function SignInSide() {
     }
   }
 
-
-  // Google Login function
-  const googleSuccess = async (res) => {
-    try {
-      const result = res.profileObj;
-      const signInfo = {
-        firstname: result.givenName.toLowerCase(),
-        lastname: result.familyName.toLowerCase(),
-        email: result.email.toLowerCase(),
-        password: res.tokenId,
-      };
-      setLoading(true);
-      try {
-        const response = await signUp(signInfo);
-        if (response.status === 201) {
-          localStorage.setItem("authToken", response.data.token);
-          window.location.href = '/';
-        }
-      } catch (e) {
-
-        const response = await login({
-          email: result.email.toLowerCase(),
-          password: res.tokenId
-        });
-
-        if (response.status === 200) {
-          localStorage.setItem('authToken', response.data.token);
-          window.location.href = '/';
-        }
-
-      }
-    } catch (e) {
-      setErrorMessage("Your Google account has been registered manually, please use email and password to log in.");
-      setTimeout(() => {
-        setErrorMessage('');
-      }, 8000);
-      setLoading(false);
-    }
-  }
-
-  const googleFailure = () => {
-    setErrorMessage("Google login was unsuccessful");
-    setTimeout(() => {
-      setErrorMessage('');
-    }, 8000);
-  }
-
   return (
     <Grid container component="main" className={classes.root}>
       <CssBaseline />
@@ -243,31 +194,7 @@ export default function SignInSide() {
               Sign in
             </Typography>
 
-            <GoogleLogin
-              clientId="804484196875-772dcm3h64clpuikj2rfkvvc5h700036.apps.googleusercontent.com"
-              render={(renderProps) => {
-                return <Button
-                  fullWidth
-                  variant="contained"
-                  color="secondary"
-                  className={classes.submit}
-                  startIcon={loading || <GoogleIcon />}
-                  onClick={renderProps.onClick}
-                  disabled={renderProps.disabled}
-                >
-                  {(loading && !loadingChecker) ? (
-                    <CircularProgress color="inherit" size="2rem" />
-                  ) : (
-                    <>Log in with Google</>
-                  )}
-                </Button>
-              }}
-              onSuccess={googleSuccess}
-              onFailure={googleFailure}
-              cookiePolicy="single_host_origin"
-            />
-
-            <Divider className={classes.divider}>OR</Divider>
+            <GoogleLoginButton setErrorMessage={setErrorMessage} setTimeout={setTimeout}/>
 
 
             {/* <form className={classes.form} noValidate> */}
